Added ability to override default Theme with a site Theme
