+++ app/components/ThemeProvider.tsx
@@ -7,17 +7,16 @@ import { useMatches, useRouteLoaderData } from 'react-router';
 import React, { useEffect, useState } from 'react';
 import { ThemeContext } from '~/hooks/useTheme';
 import useMatchesData from '~/hooks/useMatchesData';
-import { themeName, pageName, Theme, Pages } from '@/grazie';
+import { pageName, Theme, Pages } from '@/grazie';
 import { notifications } from '@mantine/notifications';
 
 export function ThemeProvider({ children }: { children: React.ReactNode }) {
   const matches = useMatches();
   const { id } = matches[matches.length - 1];
   const loader = useRouteLoaderData(id);
-  const [_theme, setTheme] = useState(themeName);
-  const [_page, setPage] = useState('root');
-  if (loader && loader?._page !== _page) {
-    setPage(loader._page);
+  const [_page, setPage] = useState(undefined);
+  if (loader?._page !== _page) {
+    setPage(loader?._page);
   }
 
   const Page = Theme?.[Pages?.[_page] ?? Theme?.[Pages?.[_page]] ?? pageName];
@@ -46,9 +45,7 @@ export function ThemeProvider({ children }: { children: React.ReactNode }) {
   }, [toast]);
 
   return (
-    <ThemeContext.Provider
-      value={{ theme: themeName, setTheme, page: pageName, setPage, data }}
-    >
+    <ThemeContext.Provider value={{ page: pageName, setPage, data }}>
       <Page>{children}</Page>
     </ThemeContext.Provider>
   );
