+++ app/theme/pages/default.tsx
@@ -1,0 +1,1 @@
+export { Page } from '~/theme/pages/Page';
