--- app/themes/README.md
@@ -1,3 +1,0 @@
-# Grazie Themes
-
-Standalone or shared themes
