--- app/themes/crimson/pages/default.tsx
@@ -1,1 +1,0 @@
-export { Page } from '~/themes/favorite/pages/Page';
