--- app/themes/favorite/styles/Footer.module.css
@@ -1,22 +1,0 @@
-.footer {
-  margin-top: rem(20px);
-  border-top: rem(1px) solid light-dark(var(--mantine-color-gray-2), var(--mantine-color-dark-5));
-}
-
-.inner {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: var(--mantine-spacing-md) var(--mantine-spacing-md);
-
-  @media (max-width: $mantine-breakpoint-sm) {
-    flex-direction: column;
-  }
-}
-
-.links {
-  @media (max-width: $mantine-breakpoint-sm) {
-    margin-top: var(--mantine-spacing-lg);
-    margin-bottom: var(--mantine-spacing-sm);
-  }
-}
