+++ grazie.config.default.ts
@@ -1,12 +1,11 @@
 /**
  * Grazie
- * @copyright Copyright (c) 2024 David Dyess II
+ * @copyright Copyright (c) 2024-2025 David Dyess II
  * @license MIT see LICENSE
  */
-import { Theme } from 'app/themes/favorite';
-export { Theme } from 'app/themes/favorite';
-// Theme name - string
-export const themeName = 'favorite';
+import { Theme } from 'app/theme';
+export { Theme } from 'app/theme';
+
 // Default page name - string
 export const pageName = 'Page';
 // Theme Object
